feat(signups): now removes prior message  when a new one is submitted (#280)
- Adds a new signup state `UPDATE_PENDING` which differentiates from
`PENDING` meaning that you may have already had a prior signup that was
approved and now your new update is pending. Also means prior `PENDING`
states get updated to `UPDATE_PENDING`.
- Checks if a `reviewMessageId` exists on a prior signup when confirming
a signup, to delete the message that had been posted
- Unsets screenshot/link if its not provided in a subsequent signup. A
signup update is assumed to be the new prog proof not cumulative to the
prior prog proof.