fix: fixes an issue with remove-signup where multiple rows were modified (#284)
There was a range error on the sheet modification for clear party
removals that was overlooked.
In addition the `/remove-signup` command has been enhanced to do the
following

- Remove a signup that matches no DB entry. - Should return error saying
this only works for signups made with the bot, not prior to the bot
because we can't verify if they're trying to remove their own or someone
elses signup
- Remove signup that matches DB entry but has not yet been approved. -
This should remove their entry from the DB but will not check the google
sheet since we don't expect to find a row there.
- Remove signup that matches DB entry and has been approved. - This
should remove their entry from the DB and check the Google Sheet for a
matching row. If a matching row exists, it will remove that row from the
Google Sheet. If no matching row exists, it will reply saying it removed
the signup from the DB but could not find a matching row in the Google
Sheet. If they think this is an error they can reach out to a
coordinator.