fix(settings): prevent prog roles from being overridden (#288)
Fixes an issue where when using `merge` with Firestore collections, an
empty object would erase the any existing properties on an object in the
document.

This is a quick fix while thinking about how to maybe restructure
elements of the collection. This fix will check if the object coming
into `upsert` is empty and assign `undefined` instead since undefined
fields will be ignored.